OptimaChatV1WritetoFile
This code will spawn a chatbot which will interact as the representative of a company and will answer user questions based on information presupplied via a business description and summary. it will write a question it does not have an answer to a predefined file. It will also record the user who is interested to get in touch to another predefined file,