OptimaChatV3AccessDBOption
This code will spawn a chat which will interact with the user as a representative of a company. It will answer questions from a predefined business description and summary which is provided to it. For any questions it cannot answer it will write it to a table of the predefined accessDB. Also it will record any user interest by writing it to another table of the accessDB. Further the chat reads a table of Questions that have been answered and appends it to its knowledge and answers the user's question which is not in the business description or summary based on this new knowledge it acquired. It checks if more questions have been answered and appends them to the knowledge in run time while the chat is ongoing.